change the dtype to float64

+++ libs/baseclass/calibrate_light.py
@@ -101,7 +101,7 @@ class CalibrateLight(Screen):
         self.ids['capture_light'].disabled = not self.ids['capture_light'].disabled
 
     def disable_clock(self):
-        self.insert_data('light', np.array(self.spec.intensities(False,True), dtype=np.float32).reshape(-1, 1))
+        self.insert_data('light', np.array(self.spec.intensities(False,True), dtype=np.float64).reshape(-1, 1))
         Clock.unschedule(self.update_graph)
     
     def on_leave(self, *args):
